requirements.txt
Django==4.0.2
djangorestframework==3.13.1
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.11.0
pytz==2021.3
requests==2.27.1
psycopg2-binary==2.9.3
pytest==7.1.0
coverage==6.3.2
python-dotenv==0.19.2